Move classes from rest api to interop
