+++ trunk/interop/src/main/java/org/mitre/mpf/interop/subject/Entity.java
@@ -24,15 +24,14 @@
  * limitations under the License.                                             *
  ******************************************************************************/
 
-package org.mitre.mpf.rest.api.subject;
+package org.mitre.mpf.interop.subject;
 
 import java.util.List;
 import java.util.Map;
 
-public record SubjectJobResult(
-        SubjectJobDetails job,
-        Map<String, String> properties,
-        Map<String, List<Entity>> entities,
-        Map<String, List<Relationship>> relationships,
-        String openmpfVersion) {
+public record Entity(
+        String id,
+        double score,
+        Map<String, List<String>> tracks,
+        Map<String, String> properties) {
 }
