+++ trunk/interop/src/main/java/org/mitre/mpf/interop/subject/SubjectJobDetails.java
@@ -24,8 +24,43 @@
  * limitations under the License.                                             *
  ******************************************************************************/
 
-package org.mitre.mpf.rest.api.subject;
+package org.mitre.mpf.interop.subject;
 
-public enum CancellationState {
-    NOT_CANCELLED, CANCELLED_BY_USER, CANCELLED_BY_SHUTDOWN
+import java.net.URI;
+import java.time.Instant;
+import java.util.Optional;
+import java.util.SortedSet;
+
+import com.fasterxml.jackson.annotation.JsonView;
+
+public record SubjectJobDetails(
+        long id,
+        SubjectJobRequest request,
+        Instant startDate,
+        Optional<Instant> endDate,
+        boolean retrievedDetectionJobs,
+        CancellationState cancellationState,
+        SortedSet<String> errors,
+        SortedSet<String> warnings,
+
+        // @JsonView is being used to prevent the "outputUri" and "callbackStatus" fields from
+        // appearing in the JSON output object.
+        @JsonView(RestView.class)
+        Optional<URI> outputUri,
+
+        // The callback occurs after the output object is written, so the callback status is not
+        // known at the time the output object is created.
+        @JsonView(RestView.class)
+        String callbackStatus) {
+
+
+    // When this view is used to instantiate a Jackson ObjectWriter with DEFAULT_VIEW_INCLUSION
+    // enabled, outputUri and callbackStatus will not be included when writing out the JSON.
+    public interface OutputObjectView {
+    }
+
+    // When this view is used, or no view is specified, all fields will be included when writing
+    // out the JSON.
+    public interface RestView {
+    }
 }
