+++ trunk/mpf-rest-api/src/main/java/org/mitre/mpf/rest/api/pipelines/Pipeline.java
@@ -33,8 +33,8 @@ import javax.validation.Valid;
 
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.constraints.NotEmpty;
+import org.mitre.mpf.interop.util.ValidName;
 import org.mitre.mpf.rest.api.util.Utils;
-import org.mitre.mpf.rest.api.util.ValidName;
 
 import com.google.common.collect.ImmutableList;
