+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/JobCompleteCallbackService.java
@@ -42,8 +42,8 @@ import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.StringEntity;
 import org.mitre.mpf.interop.JsonCallbackBody;
+import org.mitre.mpf.interop.subject.CallbackMethod;
 import org.mitre.mpf.mvc.security.OAuthClientTokenProvider;
-import org.mitre.mpf.rest.api.subject.CallbackMethod;
 import org.mitre.mpf.wfm.data.entities.persistent.BatchJob;
 import org.mitre.mpf.wfm.data.entities.persistent.DbSubjectJob;
 import org.mitre.mpf.wfm.enums.MpfConstants;
