+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/LocalStorageBackend.java
@@ -30,7 +30,7 @@ package org.mitre.mpf.wfm.service;
 import com.google.common.collect.Table;
 import org.apache.commons.lang3.mutable.Mutable;
 import org.mitre.mpf.interop.JsonOutputObject;
-import org.mitre.mpf.rest.api.subject.SubjectJobResult;
+import org.mitre.mpf.interop.subject.SubjectJobResult;
 import org.mitre.mpf.wfm.camel.operations.detection.artifactextraction.ArtifactExtractionRequest;
 import org.mitre.mpf.wfm.data.entities.persistent.BatchJob;
 import org.mitre.mpf.wfm.data.entities.persistent.MarkupResult;
