+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/component/JsonComponentDescriptor.java
@@ -34,12 +34,12 @@ import javax.validation.constraints.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
 import org.hibernate.validator.constraints.NotBlank;
+import org.mitre.mpf.interop.util.ValidName;
 import org.mitre.mpf.rest.api.pipelines.Action;
 import org.mitre.mpf.rest.api.pipelines.Algorithm;
 import org.mitre.mpf.rest.api.pipelines.Pipeline;
 import org.mitre.mpf.rest.api.pipelines.Task;
 import org.mitre.mpf.rest.api.util.MethodReturnsTrue;
-import org.mitre.mpf.rest.api.util.ValidName;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.collect.ImmutableList;
