+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/mvc/controller/TestSubjectJobController.java
@@ -55,15 +55,15 @@ import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.junit.Before;
 import org.junit.Test;
+import org.mitre.mpf.interop.subject.CallbackMethod;
+import org.mitre.mpf.interop.subject.CancellationState;
+import org.mitre.mpf.interop.subject.Entity;
+import org.mitre.mpf.interop.subject.Relationship;
+import org.mitre.mpf.interop.subject.Relationship.MediaReference;
+import org.mitre.mpf.interop.subject.SubjectJobDetails;
+import org.mitre.mpf.interop.subject.SubjectJobRequest;
+import org.mitre.mpf.interop.subject.SubjectJobResult;
 import org.mitre.mpf.interop.util.TimeUtils;
-import org.mitre.mpf.rest.api.subject.CallbackMethod;
-import org.mitre.mpf.rest.api.subject.CancellationState;
-import org.mitre.mpf.rest.api.subject.Entity;
-import org.mitre.mpf.rest.api.subject.Relationship;
-import org.mitre.mpf.rest.api.subject.Relationship.MediaReference;
-import org.mitre.mpf.rest.api.subject.SubjectJobDetails;
-import org.mitre.mpf.rest.api.subject.SubjectJobRequest;
-import org.mitre.mpf.rest.api.subject.SubjectJobResult;
 import org.mitre.mpf.test.MockitoTest;
 import org.mitre.mpf.test.TestUtil;
 import org.mitre.mpf.wfm.businessrules.SubjectJobRequestService;
