+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/wfm/businessrules/TestSubjectJobRequestService.java
@@ -58,8 +58,8 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.camel.impl.DefaultCamelContext;
 import org.junit.Test;
 import org.mitre.mpf.interop.JsonOutputObject;
-import org.mitre.mpf.rest.api.subject.CallbackMethod;
-import org.mitre.mpf.rest.api.subject.SubjectJobRequest;
+import org.mitre.mpf.interop.subject.CallbackMethod;
+import org.mitre.mpf.interop.subject.SubjectJobRequest;
 import org.mitre.mpf.test.MockitoTest;
 import org.mitre.mpf.test.TestUtil;
 import org.mitre.mpf.wfm.buffers.SubjectProtobuf;
