+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/wfm/service/TestJobCompleteCallbackService.java
@@ -29,8 +29,8 @@ import org.apache.http.message.BasicStatusLine;
 import org.junit.Before;
 import org.junit.Test;
 import org.mitre.mpf.interop.JsonCallbackBody;
+import org.mitre.mpf.interop.subject.CallbackMethod;
 import org.mitre.mpf.mvc.security.OAuthClientTokenProvider;
-import org.mitre.mpf.rest.api.subject.CallbackMethod;
 import org.mitre.mpf.test.MockitoTest;
 import org.mitre.mpf.test.TestUtil;
 import org.mitre.mpf.wfm.data.entities.persistent.BatchJob;
