+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/wfm/service/TestLocalStorageBackendImpl.java
@@ -38,8 +38,8 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
+import org.mitre.mpf.interop.subject.SubjectJobResult;
 import org.mitre.mpf.mvc.controller.TestSubjectJobController;
-import org.mitre.mpf.rest.api.subject.SubjectJobResult;
 import org.mitre.mpf.test.MockitoTest;
 import org.mitre.mpf.test.TestUtil;
 import org.mitre.mpf.wfm.data.InProgressBatchJobsService;
