+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/wfm/service/TestS3StorageBackend.java
@@ -37,11 +37,11 @@ import org.apache.commons.lang3.mutable.MutableObject;
 import org.junit.*;
 import org.junit.rules.TemporaryFolder;
 import org.mitre.mpf.interop.JsonOutputObject;
+import org.mitre.mpf.interop.subject.CancellationState;
+import org.mitre.mpf.interop.subject.SubjectJobDetails;
+import org.mitre.mpf.interop.subject.SubjectJobRequest;
+import org.mitre.mpf.interop.subject.SubjectJobResult;
 import org.mitre.mpf.rest.api.pipelines.*;
-import org.mitre.mpf.rest.api.subject.CancellationState;
-import org.mitre.mpf.rest.api.subject.SubjectJobDetails;
-import org.mitre.mpf.rest.api.subject.SubjectJobRequest;
-import org.mitre.mpf.rest.api.subject.SubjectJobResult;
 import org.mitre.mpf.test.TestUtil;
 import org.mitre.mpf.wfm.camel.operations.detection.artifactextraction.ArtifactExtractionRequest;
 import org.mitre.mpf.wfm.data.InProgressBatchJobsService;
